+++ CHANGES.rst
@@ -11,6 +11,7 @@ in progress
 - Add SQL pretty printing
 - Improve software tests
 - Unlock reading from stdin
+- Derive table name from input file name
 
 
 2023-01-xx 0.0.0